Refreshes now rotate at a adecent speed AND also work at decent speed. Score is reduced in a moderate amount. Made some corrections to the "Rahmen":  It now is displayed sans red box around it.

+++ src/global.h
@@ -81,7 +81,7 @@ EXTERN int HideInvisibleMap;
 EXTERN int PlusExtentionsOn;
 EXTERN int Alert;
 EXTERN int ThisShip;
-EXTERN long RealScore;
+EXTERN float RealScore;
 EXTERN long ShowScore;
 EXTERN long HighestScoreOfDay;
 EXTERN long LowestScoreOfDay;
